Refactored viewPenaltyDetailsService and enterPenaltyDetailsService into one service
