+++ src/test/java/uk/gov/companieshouse/web/lfp/service/lfp/impl/LFPDetailsServiceImplTest.java
@@ -6,16 +6,16 @@ import org.junit.jupiter.api.TestInstance;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.junit.jupiter.MockitoExtension;
-import uk.gov.companieshouse.web.lfp.service.lfp.EnterLFPDetailsService;
+import uk.gov.companieshouse.web.lfp.service.lfp.LFPDetailsService;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 @ExtendWith(MockitoExtension.class)
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-public class EnterLFPDetailsServiceImplTest {
+public class LFPDetailsServiceImplTest {
 
     @InjectMocks
-    private EnterLFPDetailsService enterLFPDetailsService = new EnterLFPDetailsServiceImpl();
+    private LFPDetailsService mockLFPDetailsService = new LFPDetailsServiceImpl();
 
     private static final String COMPANY_NUMBER_WITH_LETTERS = "SE123456";
 
@@ -29,7 +29,7 @@ public class EnterLFPDetailsServiceImplTest {
     @DisplayName("Append - Do not append to company number with letters")
     void validateCompanyNumberWithLettersNotAppended() throws Exception {
 
-        String companyNumber = enterLFPDetailsService.appendToCompanyNumber(COMPANY_NUMBER_WITH_LETTERS);
+        String companyNumber = mockLFPDetailsService.appendToCompanyNumber(COMPANY_NUMBER_WITH_LETTERS);
         assertEquals(companyNumber, COMPANY_NUMBER_WITH_LETTERS);
     }
 
@@ -37,7 +37,7 @@ public class EnterLFPDetailsServiceImplTest {
     @DisplayName("Append - Eight Digit company number returned the same")
     void validationEightDigitCompanyNumberReturnedTheSame() throws Exception {
 
-        String companyNumber = enterLFPDetailsService.appendToCompanyNumber(COMPANY_NUMBER_WITH_EIGHT_DIGITS);
+        String companyNumber = mockLFPDetailsService.appendToCompanyNumber(COMPANY_NUMBER_WITH_EIGHT_DIGITS);
         assertEquals(companyNumber, COMPANY_NUMBER_WITH_EIGHT_DIGITS);
     }
 
@@ -45,7 +45,7 @@ public class EnterLFPDetailsServiceImplTest {
     @DisplayName("Append - Six Digit company number should have 0's appended to beginning")
     void validationSixDigitCompanyNumberReturnedWithAppendedZeros() throws Exception {
 
-        String companyNumber = enterLFPDetailsService.appendToCompanyNumber(COMPANY_NUMBER_WITH_SIX_DIGITS);
+        String companyNumber = mockLFPDetailsService.appendToCompanyNumber(COMPANY_NUMBER_WITH_SIX_DIGITS);
         assertEquals(companyNumber, APPENDED_SIX_DIGIT_COMPANY_NUMBER);
     }
 }
